+++ Joystick/platformio.ini
@@ -8,6 +8,15 @@
 ; Please visit documentation for the other options and examples
 ; https://docs.platformio.org/page/projectconf.html
 
+[env:regular]
+platform = raspberrypi
+board = pico
+framework = arduino
+lib_deps =
+    adafruit/Adafruit ADS1X15@^2.5.0
+    SPI
+
+
 [env:debug]
 platform = raspberrypi
 board = pico
